requirements.txt
Flask==2.3.0
rembg==2.0.41
Pillow==9.3.0